# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pulse Check Backend API",
    "version": "0.1.0",
    "description": "Pulse Check backend API for entity listing, server-managed statuses, health checks, and metrics."
  },
  "paths": {
    "/entities": {
      "get": {
        "summary": "List entities",
        "operationId": "listEntities",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Hardcoded entity list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Entity"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/status-set": {
      "get": {
        "summary": "Get effective status set",
        "operationId": "getStatusSet",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Effective status set for the current user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusSet"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/statuses": {
      "get": {
        "summary": "List statuses",
        "operationId": "listStatuses",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All statuses from the effective status set.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusDefinition"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      },
      "post": {
        "summary": "Create status",
        "operationId": "createStatus",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusDefinition"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "409": {
            "description": "Status limit reached.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "status_limit_reached"
                }
              }
            }
          }
        }
      }
    },
    "/statuses/{status_id}": {
      "patch": {
        "summary": "Update status",
        "operationId": "updateStatus",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusDefinition"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      },
      "delete": {
        "summary": "Delete status",
        "operationId": "deleteStatus",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Status deleted."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Prometheus metrics",
        "operationId": "getMetrics",
        "responses": {
          "200": {
            "description": "Prometheus text metrics.",
            "content": {
              "text/plain; version=0.0.4": {
                "schema": {
                  "type": "string",
                  "example": "pulse_check_entity_list_requests_total 1\n"
                }
              }
            }
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/health/live": {
      "get": {
        "summary": "Liveness probe",
        "operationId": "getLiveness",
        "responses": {
          "200": {
            "$ref": "#/components/responses/HealthOK"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "summary": "Readiness probe",
        "operationId": "getReadiness",
        "responses": {
          "200": {
            "$ref": "#/components/responses/HealthOK"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "503": {
            "description": "Application is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                },
                "example": {
                  "status": "not_ready"
                }
              }
            }
          }
        }
      }
    },
    "/health/startup": {
      "get": {
        "summary": "Startup probe",
        "operationId": "getStartup",
        "responses": {
          "200": {
            "$ref": "#/components/responses/HealthOK"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "503": {
            "description": "Application is still starting.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                },
                "example": {
                  "status": "starting"
                }
              }
            }
          }
        }
      }
    },
    "/livez": {
      "get": {
        "summary": "Liveness probe alias",
        "operationId": "getLivenessAlias",
        "responses": {
          "200": {
            "$ref": "#/components/responses/HealthOK"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/readyz": {
      "get": {
        "summary": "Readiness probe alias",
        "operationId": "getReadinessAlias",
        "responses": {
          "200": {
            "$ref": "#/components/responses/HealthOK"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "503": {
            "description": "Application is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                },
                "example": {
                  "status": "not_ready"
                }
              }
            }
          }
        }
      }
    },
    "/startupz": {
      "get": {
        "summary": "Startup probe alias",
        "operationId": "getStartupAlias",
        "responses": {
          "200": {
            "$ref": "#/components/responses/HealthOK"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "503": {
            "description": "Application is still starting.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                },
                "example": {
                  "status": "starting"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Entity": {
        "type": "object",
        "required": [
          "id",
          "state"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "2b72d045-d9d7-43ce-9952-59a0f3e35e88"
          },
          "state": {
            "type": "string",
            "example": "active"
          }
        }
      },
      "StatusInput": {
        "type": "object",
        "required": [
          "name",
          "border_color",
          "background_color",
          "text_color"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "In progress"
          },
          "border_color": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "example": "#5e81ac"
          },
          "background_color": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "example": "#e8f0fb"
          },
          "text_color": {
            "type": "string",
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "example": "#24344d"
          }
        }
      },
      "StatusDefinition": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StatusInput"
          },
          {
            "type": "object",
            "required": [
              "id",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "example": "2b72d045-d9d7-43ce-9952-59a0f3e35e88"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "StatusSet": {
        "type": "object",
        "required": [
          "id",
          "owner_user_id",
          "role",
          "statuses"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "owner_user_id": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "status_owner",
              "participant"
            ]
          },
          "statuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatusDefinition"
            }
          }
        }
      },
      "Status": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "method_not_allowed"
          }
        }
      }
    },
    "responses": {
      "HealthOK": {
        "description": "Application health check is OK.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Status"
            },
            "example": {
              "status": "ok"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Bearer token is missing or invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "missing_bearer_token"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Request body is invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "invalid_status"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Bearer token is valid but lacks the required role.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "missing_required_role"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "status_not_found"
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "HTTP method is not allowed for this endpoint.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}